Revert "ANDROID: cpufreq: change uid_cpupower to use u32"
This reverts commit 687edd1f2b4d4e35e8cff990162aea39fba618de.

+++ drivers/cpufreq/cpufreq_stats.c
@@ -326,7 +326,7 @@ static int uid_time_in_state_show(struct seq_file *m, void *v)
 }
 
 /*
- * time_in_state is an array of u32's in the following format:
+ * time_in_state is an array of u64's in the following format:
  * [n, uid0, time0a, time0b, ..., time0n,
  *     uid1, time1a, time1b, ..., time1n,
  *     uid2, time2a, time2b, ..., time2n, etc.]
@@ -336,8 +336,8 @@ static int time_in_state_show(struct seq_file *m, void *v)
 {
 	struct uid_entry *uid_entry;
 	unsigned long bkt;
-	u32 cpufreq_max_states = all_freq_table->table_size;
-	u32 uid, time;
+	u64 cpufreq_max_states = all_freq_table->table_size;
+	u64 uid, time;
 	int i;
 
 	if (!cpufreq_all_freq_init)
@@ -356,12 +356,12 @@ static int time_in_state_show(struct seq_file *m, void *v)
 			max_state = uid_entry->alive_max_states;
 
 		if (max_state) {
-			uid = (u32) uid_entry->uid;
+			uid = (u64) uid_entry->uid;
 			seq_write(m, &uid, sizeof(uid));
 		}
 
 		for (i = 0; i < max_state; ++i) {
-			u32 total_time_in_state = 0;
+			u64 total_time_in_state = 0;
 
 			if (uid_entry->dead_time_in_state &&
 				i < uid_entry->dead_max_states) {
@@ -373,7 +373,7 @@ static int time_in_state_show(struct seq_file *m, void *v)
 				total_time_in_state +=
 					uid_entry->alive_time_in_state[i];
 			}
-			time = (u32)
+			time = (u64)
 				cputime_to_clock_t(total_time_in_state);
 			seq_write(m, &time, sizeof(time));
 		}
@@ -388,7 +388,7 @@ static int time_in_state_show(struct seq_file *m, void *v)
 }
 
 /*
- * concurrent_active_time is an array of u32's in the following format:
+ * concurrent_active_time is an array of u64's in the following format:
  * [n, uid0, time0a, time0b, ..., time0n,
  *     uid1, time1a, time1b, ..., time1n,
  *     uid2, time2a, time2b, ..., time2n, etc.]
@@ -398,7 +398,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 {
 	struct uid_entry *uid_entry;
 	struct task_struct *task, *temp;
-	u32 uid, time, num_possible_cpus = num_possible_cpus();
+	u64 uid, time, num_possible_cpus = num_possible_cpus();
 	unsigned long bkt, flags;
 	int i;
 
@@ -442,11 +442,11 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 			!uid_entry->dead_concurrent_active_time)
 			continue;
 
-		uid = (u32) uid_entry->uid;
+		uid = (u64) uid_entry->uid;
 		seq_write(m, &uid, sizeof(uid));
 
 		for (i = 0; i < num_possible_cpus; ++i) {
-			u32 total_concurrent_active_time = 0;
+			u64 total_concurrent_active_time = 0;
 
 			if (uid_entry->dead_concurrent_active_time)
 				total_concurrent_active_time =
@@ -454,7 +454,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 			if (uid_entry->alive_concurrent_active_time)
 				total_concurrent_active_time +=
 					uid_entry->alive_concurrent_active_time[i];
-			time = (u32)
+			time = (u64)
 				cputime_to_clock_t(total_concurrent_active_time);
 			seq_write(m, &time, sizeof(time));
 		}
@@ -470,7 +470,7 @@ static int concurrent_active_time_show(struct seq_file *m, void *v)
 }
 
 /*
- * concurrent_policy_time is an array of u32's in the following format:
+ * concurrent_policy_time is an array of u64's in the following format:
  * [n, x0, ..., xn, uid0, time0a, time0b, ..., time0n,
  *                  uid1, time1a, time1b, ..., time1n,
  *                  uid2, time2a, time2b, ..., time2n, etc.]
@@ -483,8 +483,8 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 	struct task_struct *task, *temp;
 	struct cpufreq_policy *policy;
 	struct cpufreq_policy *last_policy = NULL;
- 	u32 buf[num_possible_cpus()];
-	u32 uid, time;
+ 	u64 buf[num_possible_cpus()];
+	u64 uid, time;
 	unsigned long bkt, flags;
 	int i, cnt = 0, num_possible_cpus = num_possible_cpus();
 
@@ -510,7 +510,7 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 	if (last_policy)
 		cpufreq_cpu_put(last_policy);
 
- 	buf[0] = (u32) cnt;
+ 	buf[0] = (u64) cnt;
 	seq_write(m, buf, (cnt + 1) * sizeof(*buf)); 
 
 	rt_mutex_lock(&uid_lock);
@@ -548,11 +548,11 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 			!uid_entry->dead_concurrent_policy_time)
 			continue;
 
- 		uid = (u32) uid_entry->uid;
+ 		uid = (u64) uid_entry->uid;
 		seq_write(m, &uid, sizeof(uid));
 
 		for (i = 0; i < num_possible_cpus; ++i) {
-			u32 total_concurrent_policy_time = 0;
+			u64 total_concurrent_policy_time = 0;
 
 			if (uid_entry->dead_concurrent_policy_time)
 				total_concurrent_policy_time =
@@ -560,7 +560,7 @@ static int concurrent_policy_time_show(struct seq_file *m, void *v)
 			if (uid_entry->alive_concurrent_policy_time)
 				total_concurrent_policy_time +=
 					uid_entry->alive_concurrent_policy_time[i];
- 			ime = (u32)
+ 			ime = (u64)
 				cputime_to_clock_t(total_concurrent_policy_time);
 			seq_write(m, &time, sizeof(time));
 		}
